Fix value too long error, update schema, and improve Resume API response mapping

+++ src/main/java/com/sarangsvkm/portfolio_api/repository/EducationRepository.java
@@ -7,4 +7,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.sarangsvkm.portfolio_api.entity.*;
 
 
-public interface EducationRepository extends JpaRepository<Education, Long> {}
+public interface EducationRepository extends JpaRepository<Education, Integer> {}
