+++ src/main/java/com/sarangsvkm/portfolio_api/repository/ExperienceRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.sarangsvkm.portfolio_api.entity.Experience;
 
 
-public interface ExperienceRepository extends JpaRepository<Experience, Long> {}
+public interface ExperienceRepository extends JpaRepository<Experience, Integer> {}
