+++ src/main/java/com/sarangsvkm/portfolio_api/repository/ProfileRepository.java
@@ -7,4 +7,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.sarangsvkm.portfolio_api.entity.Profile;
 
 
-public interface ProfileRepository extends JpaRepository<Profile, Long> {}
+public interface ProfileRepository extends JpaRepository<Profile, Integer> {}
