+++ src/main/java/com/sarangsvkm/portfolio_api/repository/ProjectRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sarangsvkm.portfolio_api.entity.Project;
 
-public interface ProjectRepository extends JpaRepository<Project, Long> {}
+public interface ProjectRepository extends JpaRepository<Project, Integer> {}
